requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
python-dotenv
pymupdf
pytest
playwright
jinja2
langchain-core